requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
streamlit==1.26.0
openai==1.7.2
python-dotenv==1.0.0
aiohttp==3.8.5
firecrawl==0.1.5
serpapi==0.1.0
requests==2.31.0
gunicorn==21.2.0
plotly==5.17.0
pandas==2.1.0
matplotlib==3.8.0
pydantic==2.0.3 